Now depending on java-assist and the guava-base
Future commits might change the whole structure of api / internals again.
Im thinking about how to provide a whole in one thingy without having
to make everyone depend on java-assist who is using Icarus.
Moving the api to another module again would'nt help though.
The Internals are needed in the classpath in order for the whole thing to work.