+ Added a really basic and tiny Factory for caches.
This already shows some things planned for the caches, every cache
should use the same executor in order to be able to create a bunch
of caches without having thousands of threads running. The factory will
later get some more attributes that can be set but that's it for now.